+++ samples/publish-to-github-packages/src/main/kotlin/index.kt
@@ -1,0 +1,18 @@
+package sandbox
+
+import io.ktor.client.HttpClient
+
+external interface GreetingArgs {
+  val name: String
+  val sureName: String
+}
+
+@JsExport
+fun buildGreeting(args: GreetingArgs): String {
+  return "Hi ${args.name} ${args.sureName}! Here's HttpClient::class.simpleName: ${HttpClient::class.simpleName}"
+}
+
+@JsExport
+fun greet(args: GreetingArgs) {
+  println(buildGreeting(args))
+}